Run copy elimination and field pruning on consolidated rectangles
